SDLGui: Method pixelWall to add a wall
This method paint a pixel in the color of the walls. The pixels painted
this way are kept in an SDL_Surface imageWalls with a "color key" to
make it transparent so that it can be blitted on top of other surfaces.

The surface imageWalls is blitted on top of scratch and on top of
imageBin when the reset operation is required.

Signed-off-by: SylvainGault <github@sylvaingault.fr>