SDLGui: Merge mouse motion events
When the processing is too slow compared to the rate of the mouse motion
events, the events stack up until the buffer is full and this create a
long latency between mouse motion and drawing.

So the current mouse motion event is merged with the next one if it is
compatible.

Signed-off-by: SylvainGault <github@sylvaingault.fr>